# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for processing bracketed images with AutoEnhance AI HDR merging. This API handles order creation, image uploads, HDR processing, and real-time status updates via Supabase Realtime.",
        "title": "Instant HDR Backend API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/orders/{order_id}/upload": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Uploads multiple bracketed images to an AutoEnhance AI order.\n\n**Automatic Grouping (Default):**\n- All images in one upload call are automatically assigned the same group UUID\n- This makes each upload = one HDR image\n- Example: Upload 3 brackets → They all get the same group_id → Process as 1 HDR\n\n**Custom Grouping (Advanced):**\n- Optionally specify different group IDs for files in the same upload\n- Example: groups=\"living-room,living-room,living-room,kitchen,kitchen,kitchen\"\n- This creates multiple HDR groups in one upload call\n\n**Workflow:**\n1. Upload bedroom brackets (3 images) → Auto-grouped as one HDR\n2. Upload kitchen brackets (3 images) → Auto-grouped as another HDR\n3. Process with bracket_grouping=\"by_upload_group\" → 2 HDR images",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "Upload images with automatic or custom grouping",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID (UUID)",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Bracketed images (multiple files allowed)",
                        "name": "images",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Advanced: Custom group ID for each file (comma-separated). If not provided, all files get the same auto-generated UUID.",
                        "name": "groups",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.FileInfo": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "models.UploadErrorInfo": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "stage": {
                    "description": "\"create_bracket\", \"upload\", \"verify\", \"database\"",
                    "type": "string"
                }
            }
        },
        "models.UploadResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UploadErrorInfo"
                    }
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FileInfo"
                    }
                },
                "order_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}